requirements.txt
atomicwrites==1.4.1
attrs==22.1.0
autopep8==1.7.0
bcrypt==4.0.0
bleach==5.0.1
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
cryptography==38.0.1
distlib==0.3.6
docutils==0.19
fabric==2.7.1
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
invoke==1.7.1
itsdangerous==2.1.2
jaraco.classes==3.2.2
Jinja2==3.1.2
keyring==23.9.1
Kivy==2.1.0
Kivy-examples==2.1.0
Kivy-Garden==0.1.5
MarkupSafe==2.1.1
metamaterial-generator @ file:///Users/jmankoff/Research/embroidery/meta-embroidery/dist/metamaterial_generator-0.1.0-py3-none-any.whl
more-itertools==8.14.0
nodeenv==1.7.0
numpy==1.23.3
packaging==21.3
paramiko==2.11.0
parcel==0.2.2
pathlib2==2.3.7.post1
Pillow==9.2.0
pkginfo==1.8.3
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.13.0
PyNaCl==1.5.0
pyparsing==3.0.9
pytest==7.1.3
pytest-runner==6.0.0
readme-renderer==37.1
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.5.1
scipy==1.9.1
six==1.16.0
SQLAlchemy==1.4.41
svgpathtools==1.5.1
svgwrite==1.4.3
toml==0.10.2
tomli==2.0.1
twine==4.0.1
urllib3==1.26.12
virtualenv==20.16.5
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
zipp==3.8.1
